pyproject.toml
[tool.poetry]
name = "hs-formation"
version = "4.1.13"
authors = ["Dotan Nahum <jondotan@gmail.com>", "HiredScoreLabs <oss@hiredscore.com>"]
description = "A generic functional middleware infrastructure for Python."

license = "MIT"

readme = "README.md"

homepage = "https://github.com/HiredScorelabs/hs-formation"
repository = "https://github.com/HiredScorelabs/hs-formation"
documentation = "https://github.com/jondot/formation"

keywords = ["flask", "django", "requests", "middleware", "functional"]

[tool.poetry.dependencies]

python = ">=3.7"
toolz = "^0.9.0"
cytoolz = "^0.9.0"
requests = "^2.20"
pybreaker = "^0.4.5"
xmltodict = "^0.11.0"
lxml = "^4.2"
attrs = "^19.1"
attrs-serde = "^0.2.2"
aiohttp = "=^3.7.3"
aiobreaker = "=^1.1.0"

[tool.poetry.dev-dependencies]

pytest-cov = "^2.6"
pytest-sugar = "^0.9.2"
boltons = "^18.0"
pytest = "^6.2.1"
pytest-benchmark = "^3.1"
pytest-asyncio = "=^0.14.0"
snapshottest =  { git = "https://github.com/jondot/snapshottest", branch="serialize-snapshots" }

flake8 = { version = "^3.5" }
requests = "^2.20"
pytest-vcr = "^1.0"
structlog = "^18.2"

[tool.poetry.extras]
lint = ["flake8-mypy"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
